Update deployment.mdx
Having two separate test and sync workflows made it so they a sync action couldn't easily be dependant on a test actions. They were also running on any changes to the repo, rather that only the detections directory. Wrote a combined workflow to simplify this and prevent excess test and sync actions.